don't manage the app context in Jenkins config
Prior to this commit, we appended the application context to the
jenkinsURL in Jenkins' location config and the Mesos configuration in
config.xml.

By running Jenkins in Tomcat, we're able to use the root (/) path and
simply have Tomcat do the re-write, so that we can run this service
behind the reverse proxy we provide users in DCOS. Therefore, appending
the context here is unnecessary.